beam lerp and score display
changed the movement of the beam to lerp so the balls don't go under the beam when rotating.
score is displayed for each coloured box, and the total score is displayed at the side. (black balls empty the score of the colour they fall in, the total score is then again calculated)